Refactor command handler and error handling
Move command handler to `utils` and add error handling decorator.

* **Command Handler:**
  - Add `utils/command_handler.py` to handle commands.
  - Move command handler logic from `selfbot.py` to `utils/command_handler.py`.
  - Define `CommandHandler` class to handle commands and interactions.

* **Error Handling:**
  - Add `utils/error_handler.py` to handle errors.
  - Define `error_handler` decorator to handle errors.
  - Refactor existing error handling logic into the decorator.

* **SelfBot:**
  - Remove command handler logic and import from `utils/command_handler.py`.
  - Remove error handling logic and import from `utils/error_handler.py`.
  - Update `SelfBot` class to use `CommandHandler` and `error_handler`.

* **Cogs:**
  - Remove `try-except` blocks and use `error_handler` decorator in `cogs/custom_activity_cog.py`, `cogs/help_cog.py`, `cogs/moderation_cog.py`, `cogs/name_history_cog.py`, and `cogs/avatar_history_cog.py`.
  - Import `error_handler` from `utils/error_handler.py` in each cog file.
  - Apply `error_handler` decorator to command methods in each cog file.
