+++ cogs/help_cog.py
@@ -1,11 +1,13 @@
 import discord
 from discord.ext import commands
+from utils.error_handler import error_handler
 
 class HelpCog(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
 
     @commands.command(name='help')
+    @error_handler
     async def help_command(self, ctx, *input):
         prefix = self.bot.command_prefix
         version = "1.0"
